requirements.txt
openai>=1.14.3
streamlit>=1.32.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0 